+++ CHANGES.md
@@ -1,5 +1,6 @@
 # Changes
 
+* 2019-10-02 2.0.0 Update for Rails 6 & Minitest 6
 * 2018-06-21 1.3.1 Update SimpleCov syntax; drop Coveralls
 * 2017-11-27 1.3.0 Add messages with better test coverage
 * 2017-11-21 1.2.0 Add tests for nil and each for actual value